BigUpdate
Removed all bugs from alot of functions, for atleast the first month of the game / sim.
Added a bed with the ability to end the day.
Used the phone to tie a function to order food instead of making food.
Cleaned out some code, but also added a lot more constraints to code and comments as well as new functions so the code grew by about 100 lines.